feat(datastructs): Added in a Dynamic Array
Introduced a Dynamic Array into the datastructs directory of the stdlib. This was done as it was needed to continue completion of the Log.

I will be commiting along with this the work I have made on the log. The log at a base level is functional, but I have decided to implement some changes to help with performance. This is why I needed a Dynamic Array.