docs(dynamic array): added docs for dynamic array
Added in Doxygen documentation for the jester-dynamic-array.h file, allowing for easier use.
Added in inline documentation to the jester-dynamic-array.c file. this will help if me or someone else needs to fix or upgrade it at some point down the line.
A few of the functions got a rework. this was mostly a change of order to make sure they fail out before doing anything too serious.